+++ HISTORY.md
@@ -3,10 +3,11 @@
 ## vNEXT
 
 * Added support for specifying the service(s) to start/stop.
+* Added support for docker-compose's `--tls`* parameters.
 
 ## v0.1.1
 
-* Added support for docker-compose's --remove-orphans parameter.
+* Added support for docker-compose's `--remove-orphans` parameter.
 * The cookbook is now compatible with CentOS.
 * Improved documentation.